The Kettenkrad: A Motorcycle Half-Track
The Kettenkrad is a half-motorcycle, half-tracked vehicle used by the German Army in World War II. This 2-person collaborative project took me around 150 hours to create. Includes working mechanics, such as the tracks in gear with all wheels, rotation of the front wheel by the handlebars, and several levers in the driver's seat.